assignment-2-784478/code/clientstreamingestapp/requirements.txt
paho-mqtt
pymongo
apscheduler
